1318 - move recurring-donation folder to admin
