+++ src/components/admin/recurring-donation/Form.tsx
@@ -18,7 +18,7 @@ import { endpoints } from 'service/apiEndpoints'
 import { AlertStore } from 'stores/AlertStore'
 import FormTextField from 'components/common/form/FormTextField'
 import SubmitButton from 'components/common/form/SubmitButton'
-import PersonSelectDialog from 'components/person/PersonSelectDialog'
+import PersonSelectDialog from 'components/admin/person/PersonSelectDialog'
 import { Form, Formik } from 'formik'
 import { RecurringDonationStatus } from 'gql/recurring-donation-status.d'
 import { PersonResponse } from 'gql/person'
